Good meme
This reverts commit 911f90cef3d3c668f094d865db823327fd12ef8f.

+++ ftc_app/TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Bot.java
@@ -857,7 +857,7 @@ public class Bot {
 
     }
 
-    public void strafeAdjusts(double targetHeading, MovementEnum direction) {
+    public void strafeAdjusts(int targetHeading, MovementEnum direction) {
         double headingError = targetHeading + imu.getAngularOrientation().firstAngle;
         if (Math.abs(headingError) < 1) {
             FL.setPower(FR.getPower());
